pyproject.toml
[tool.poetry]
name = "cloud-tech-webapp"
version = "0.1.0"
description = ""
authors = ["ishan-modi <ishan.modi@glib.ai>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
django-tenants = "^3.5.0"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
python-environ = "^0.4.54"
markdown = "^3.4.3"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
